Module trezor_crypto_lib::ed25519::sha3

ed25519 API using sha3 signatures, equivalent to ed25519-donna APIs generated with a custom sha3 hasher (see tests/ed25519-sha3.c)

Functions

dalek_ed25519_publickey_sha3

Derives a public key from a private key using sha3 digest

dalek_ed25519_sign_open_sha3

Verifies a message using the provided secret key using sha3 digest

dalek_ed25519_sign_sha3

Signs a message using the provided secret key using sha3 digest